requirements.txt
numpy<2.0.0
faiss-cpu>=1.8.0
openai>=1.0.0
streamlit>=1.0.0
python-dotenv
pydantic-settings>=2.10.0
sentence-transformers>=2.5
neo4j
tiktoken==0.7.0
flake8
pytest
pyyaml
jsondiff
jsonschema